# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nova Launch API",
    "version": "1.0.0",
    "description": "REST API for the Nova Launch Stellar token deployment platform.",
    "contact": {
      "name": "Nova Launch",
      "url": "https://github.com/Emmyt24/Nova-launch"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "https://api.nova-launch.app",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/api/vaults/creator/{address}": {
      "get": {
        "tags": [
          "Vaults"
        ],
        "summary": "Vaults created by address",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vault list"
          }
        }
      }
    },
    "/api/vaults/beneficiary/{address}": {
      "get": {
        "tags": [
          "Vaults"
        ],
        "summary": "Vaults where address is the beneficiary",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vault list"
          }
        }
      }
    },
    "/api/vaults/{id}": {
      "get": {
        "tags": [
          "Vaults"
        ],
        "summary": "Get vault by on-chain ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vault detail"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from POST /auth/verify or POST /auth/refresh"
      }
    },
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Admin JWT issued for users with role 'admin' or 'super_admin'."
      }
    },
...
  }
}